chore: move tests to tests
