--- examples/schemas/style-options.schema.json
@@ -1,28 +1,0 @@
-{
-  "$ref": "$defs/style-options",
-  "$defs": {
-    "style-options": {
-      "type": "object",
-      "properties": {
-        "color": {
-          "type": "string"
-        },
-        "styleCodes": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          }
-        },
-        "htmlStyleFormat": {
-          "type": "string",
-          "enum": ["css", "className"],
-          "description": "HTML style format"
-        }
-      },
-      "required": ["color"],
-      "additionalProperties": false
-    }
-  },
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "version": "0.1.1"
-}
