--- examples/schemas/token.schema.json
@@ -1,47 +1,0 @@
-{
-  "$ref": "$defs/token",
-  "$defs": {
-    "token": {
-      "type": "object",
-      "properties": {
-        "content": {
-          "type": "string",
-          "description": "The actual text content of the token"
-        },
-        "metadata": {
-          "type": "object",
-          "properties": {
-            "style": {
-              "type": "object",
-              "properties": {
-                "color": {
-                  "type": "string"
-                },
-                "styleCodes": {
-                  "type": "array",
-                  "items": {
-                    "type": "string"
-                  }
-                },
-                "htmlStyleFormat": {
-                  "type": "string",
-                  "enum": ["css", "className"],
-                  "description": "HTML style format"
-                }
-              },
-              "required": ["color"],
-              "additionalProperties": false,
-              "description": "Styling information for this token"
-            }
-          },
-          "additionalProperties": {},
-          "description": "Additional token metadata including style information"
-        }
-      },
-      "required": ["content"],
-      "additionalProperties": false
-    }
-  },
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "version": "0.1.1"
-}
