--- tests/manual/test-cli.ts
@@ -1,4 +1,0 @@
-#!/usr/bin/env bun
-
-// Test the new CLI directly without building
-import "../src/cli/cli";
